app/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
protobuf==4.21.12
pylogix==0.8.14
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
zipp==3.16.2
gunicorn ==21.2.0
retrying==1.3.4
pika==1.3.2